Testing: Move lsp_manager fixture to the LSP tests
- Also add it to conftest.py to avoid importing it everywhere.
